feat: Goal and CheckIn pages share the same Targets field
